Reorganise integration tests
* Renames integration test filenames so that they have a `_test.rs`
  suffix, to prevent cases where the implementation and its test have
  the same filename, making it hard to distinguish between editor tabs
  in an IDE.
* Moves the integration tests from `tests/integration/` to `tests/` since
  the additional nesting was unnecessary (this also required disabling
  automatic integration test discovery via `autouse = false`, now that
  `tests/mod.rs` exists and would otherwise get imported twice).
* Renames the `integration_tests` module to `tests` so that it matches
   the directory name, to reduce the potential for confusion.

See:
https://doc.rust-lang.org/cargo/guide/project-layout.html
https://doc.rust-lang.org/cargo/reference/cargo-targets.html#target-auto-discovery
https://doc.rust-lang.org/reference/items/modules.html#the-path-attribute

And for general background on testing strategies/layouts:
https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html
http://xion.io/post/code/rust-unit-test-placement.html
https://joshleeb.com/posts/rust-integration-tests.html

GUS-W-14069572.
